requirements.txt
requests
bs4
datetime
markupsafe
Flask == 2.3.1
pandas
selenium == 4.15.2
streamlit_aggrid
gunicorn==20.0.4
python-dotenv
sqlalchemy
mysqlcient